requirements.txt
PyMuPDF
Pillow
opencv-python
tensorflow-gpu==1.14
